*** empty log message ***

+++ realtime/classes/dbrealtime_class_inc.php
@@ -1,10 +1,13 @@
 <?php
 
-/*
+
+/**
  * Created on 15 Mar 2007
  * @author Mohamed Yusuf
  */
-// security check - must be included in all scripts
+/**
+ * security check - must be included in all scripts
+ */
 if (!$GLOBALS['kewl_entry_point_run'])
 {
 	die("You cannot view this page directly");
@@ -21,7 +24,6 @@ class dbrealtime extends dbTable
 	}
 
 	/**
-	 * @author Mohamed Yusuf
 	 * @param string userId user identification number
 	 * @param string userLevel group that user belongs to
 	 * @param string contextCode Course code
@@ -29,26 +31,36 @@ class dbrealtime extends dbTable
 	 */
 	public function requestToken($userId, $userLevel, $contextCode)
 	{
-		
+		//log_debug("userid: " . $userId . " userlevel: " . $userLevel . " contextcode: " . $contextCode);
 		$userLevel = strtolower($userLevel);
 		$convId = $this->getConversationId($contextCode);
-		$sql = "select hastoken from tbl_realtime_voicequeue where conversationid=\"$convId\"";
+		$sql = "select hastoken, puid from tbl_realtime_voicequeue where conversationid=\"$convId\"";
 		$result = $this->getArray($sql);
-		if (eregi("student",$userLevel))
+
+		foreach ($result as $res)
 		{
-			if(empty($result)){
+			$hasToken = $res['hastoken'];
+			$puid = $res['puid'];
+		}
+		if (eregi("student", $userLevel) || eregi("guest", $userLevel) || eregi("administrator", $userLevel))
+		{
+			if (empty ($result))
+			{
 				$this->addQueue($userId, $userLevel, $contextCode);
-				return SUCCESS;				
-			} else { 
-				foreach ($result as $res)
-				{
-					$hasToken = $res['hastoken'];
-				}
+				return SUCCESS;
+			} else
+			{
 				if ("1" == $hasToken)
 				{
+					$inQueue = array ();
+					$inQueue['conversationid'] = $this->getConversationId($contextCode);
+					$inQueue['userid'] = $userId;
+					$inQueue['hastoken'] = 0;
+					$this->insert($inQueue, 'tbl_realtime_voicequeue');
 					return FAILURE;
-				} else {
-					//assumed that token had been released
+				} else
+				{
+					//assumed that token have been released
 					$this->addQueue($userId, $userLevel, $contextCode);
 					return SUCCESS;
 				}
@@ -57,7 +69,8 @@ class dbrealtime extends dbTable
 		elseif (eregi("lecturer", $userLevel))
 		{
 			//explicitly realease token
-			$this->releaseToken($userId, $userLevel, $contextCode);
+			$sql = "delete from tbl_realtime_voicequeue where hastoken=1";
+			$this->getArray($sql);
 			$this->addQueue($userId, $userLevel, $contextCode);
 			return SUCCESS;
 		} else
@@ -68,20 +81,34 @@ class dbrealtime extends dbTable
 	}
 
 	/**
-	* @author Mohamed Yusuf
-	* @param string contextCode Course code
-	* @return String SUCCESS/FAILURE
-	*/
+	 * @param string contextCode Course code
+	 * @return String SUCCESS/FAILURE
+	 */
 	public function releaseToken($userId, $userLevel, $contextCode)
 	{
 		$convId = $this->getConversationId($contextCode);
-		$sql = "delete from tbl_realtime_voicequeue where conversationid=\"$convId\"";
+		$sql = "delete from tbl_realtime_voicequeue where conversationid=\"$convId\" && userid=\"$userId\"";
 		$this->getArray($sql);
+
+		$sql = "select puid from tbl_realtime_voicequeue where conversationid=\"$convId\" order by puid limit 1";
+		$result = $this->getArray($sql);
+		if (empty ($result))
+		{
+			//do nothing
+		} else
+		{
+			foreach ($result as $res)
+			{
+				$puid = $res['puid'];
+			}
+			$update = array ();
+			$update['hastoken'] = 1;
+			$this->update("puid", $puid, $update, "tbl_realtime_voicequeue");
+		}
 		return SUCCESS;
 	}
 
 	/**
-	 * @author Mohamed Yusuf
 	 * @param string contextCode Course code
 	 * @return String SUCCESS/FAILURE
 	 */
@@ -100,18 +127,17 @@ class dbrealtime extends dbTable
 	}
 
 	/**
-	* @author Mohamed Yusuf
-	* @param string userId user identification number
-	* @param string userLevel group that user belongs to
-	* @param string contextCode Course code
-	* @return String SUCCESS/FAILURE
-	*/
+	 * @param string userId user identification number
+	 * @param string userLevel group that user belongs to
+	 * @param string contextCode Course code
+	 * @return String SUCCESS/FAILURE
+	 */
 	public function startConversation($userId, $userLevel, $contextCode)
 	{
 		$sql = "select contextcode from tbl_realtime_conversation where contextcode=\"$contextCode\"";
 		$result = $this->getArray($sql);
 
-		if (empty($result))
+		if (empty ($result))
 		{
 			$this->createConversation($userId, $userLevel, $contextCode);
 			$this->addQueue($userId, $userLevel, $contextCode);
@@ -126,10 +152,68 @@ class dbrealtime extends dbTable
 	}
 
 	/**
-	* @author Mohamed Yusuf
-	* @param string contextCode Course code
-	* @return String SUCCESS/FAILURE
-	*/
+	 * @param string userId user identification number
+	 * @param string userLevel group that user belongs to
+	 * @param string contextCode Course code
+	 * @return String SUCCESS/FAILURE
+	 */
+	public function joinConversation($userId, $userLevel, $contextCode)
+	{
+		$convId = $this->getConversationId($contextCode);
+		if (empty ($convId))
+		{
+			return FAILURE;
+		} else
+		{
+			return SUCCESS;
+		}
+	}
+
+	/**
+	 * @param string userId user identification number
+	 * @param string userLevel group that user belongs to
+	 * @param string contextCode Course code
+	 * @return String SUCCESS/FAILURE
+	 */
+	public function leaveConversation($userId, $userLevel, $contextCode)
+	{
+		return SUCCESS;
+	}
+
+	/**
+	 * @param string userId user identification number
+	 * @param string userLevel group that user belongs to
+	 * @param string contextCode Course code
+	 * @return String SUCCESS/FAILURE
+	 */
+	public function checkToken($userId, $userLevel, $contextCode)
+	{
+		$sql = "select hastoken from tbl_realtime_voicequeue where userid=\"$userId\"";
+		$result = $this->getArray($sql);
+		if (empty ($result))
+		{
+			//log_debug("empty array, statement did not fetch rows");
+			return "false";
+		} else
+		{
+			foreach ($result as $res)
+			{
+				$hasToken = $res['hastoken'];
+			}
+			if (1 == $hasToken)
+			{
+				return "true";
+			} else
+			{
+				return "false";
+			}
+		}
+	}
+
+	/**
+	 * @param string contextCode Course code
+	 * @return String SUCCESS/FAILURE
+	 */
 	private function removeConversation($contextCode)
 	{
 		$convId = $this->getConversationId($contextCode);
@@ -144,12 +228,11 @@ class dbrealtime extends dbTable
 	}
 
 	/**
-	* @author Mohamed Yusuf
-	* @param string userId user identification number
-	* @param string userLevel group that user belongs to
-	* @param string contextCode Course code
-	* @return String SUCCESS/FAILURE
-	*/
+	 * @param string userId user identification number
+	 * @param string userLevel group that user belongs to
+	 * @param string contextCode Course code
+	 * @return String SUCCESS/FAILURE
+	 */
 	private function addQueue($userId, $userLevel, $contextCode)
 	{
 		$lastId = $this->getConversationId($contextCode);
@@ -161,10 +244,9 @@ class dbrealtime extends dbTable
 	}
 
 	/**
-	* @author Mohamed Yusuf
-	* @param string contextCode Course code
-	* @return String id
-	*/
+	 * @param string contextCode Course code
+	 * @return String id
+	 */
 	private function getConversationId($contextCode)
 	{
 		$sql = "select id  from tbl_realtime_conversation where contextcode=\"$contextCode\"";
@@ -177,12 +259,11 @@ class dbrealtime extends dbTable
 	}
 
 	/**
-	* @author Mohamed Yusuf
-	* @param string userId user identification number
-	* @param string userLevel group that user belongs to
-	* @param string contextCode Course code
-	* @return String SUCCESS/FAILURE
-	*/
+	 * @param string userId user identification number
+	 * @param string userLevel group that user belongs to
+	 * @param string contextCode Course code
+	 * @return String SUCCESS/FAILURE
+	 */
 	private function createConversation($userId, $userLevel, $contextCode)
 	{
 		$addRow = array ();
@@ -191,58 +272,6 @@ class dbrealtime extends dbTable
 		$addRow['contextcode'] = $contextCode;
 		$addRow['starttime'] = $starttime;
 		$this->insert($addRow, 'tbl_realtime_conversation');
-	}
-
-	/**
-	* @author Mohamed Yusuf
-	* @param string userId user identification number
-	* @param string userLevel group that user belongs to
-	* @param string contextCode Course code
-	* @return String SUCCESS/FAILURE
-	*/
-	public function joinConversation($userId, $userLevel, $contextCode)
-	{
-		$convId = $this->getConversationId($contextCode);
-		if (empty ($convId))
-		{
-			return FAILURE;
-		} else
-		{
-			return SUCCESS;
-		}
-	}
-
-	/**
-	* @author Mohamed Yusuf
-	* @param string userId user identification number
-	* @param string userLevel group that user belongs to
-	* @param string contextCode Course code
-	* @return String SUCCESS/FAILURE
-	*/
-	public function leaveConversation($userId, $userLevel, $contextCode)
-	{
-		return SUCCESS;
-	}
-
-	/**
-	* @author Mohamed Yusuf
-	* @param string userId user identification number
-	* @param string userLevel group that user belongs to
-	* @param string contextCode Course code
-	* @return String SUCCESS/FAILURE
-	*/	
-	public function checkToken($userId, $userLevel, $contextCode)
-	{
-		$sql = "select hastoken from tbl_realtime_voicequeue where userid=\"$userId\"";
-		$result = $this->getArray($sql);
-		if(empty($result)){
-			//log_debug("empty array, statement did not fetch rows");
-			return FAILURE;
-		}else{
-			//log_debug("statement did fetch rows");
-			return SUCCESS;
-		}
-		
 	}
 }
 ?>
